% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fancyNumbers.R
\name{fancyNumbers}
\alias{fancyNumbers}
\title{Creates pleasing formatting of 10^x log scale}
\usage{
fancyNumbers(n)
}
\arguments{
\item{n}{one or a list of numbers that are a factor of 10}
}
\value{
scientific notation format of n
}
\description{
Takes a single or list of numbers that are a factor of 10 and
formats them in 10^x notation. A great use case for this
would be formatting of axis labels
}
\examples{
fancyNumbers(10)
fancyNumbers(c(1,10,100,1000))

}
