% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipagePDFsave.R
\name{multipagePDFsave}
\alias{multipagePDFsave}
\title{Multipage .pdf wrapper function}
\usage{
multipagePDFsave(ggplot_list, pdf_name, page_width = 11, page_height = 8.5)
}
\arguments{
\item{ggplot_list}{a list with each list element containing a ggplot object}

\item{pdf_name}{a text string filename ending in .pdf}

\item{page_width}{numeric value of desired page width.
Default value of 11 provided}

\item{page_height}{numeric value of desired page height
Default value of 8.5 provided}
}
\value{
writes .pdf file with one ggplot object per page
}
\description{
Save several ggplots collected in a list to a single multipage pdf
}
