% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettyLogs.R
\name{prettyLogs}
\alias{prettyLogs}
\title{ggplot helper function}
\usage{
prettyLogs(x)
}
\arguments{
\item{x}{inherits ggplot object
e.g. scale_y_continuous(breaks=prettyLogs)}
}
\value{
pretty_logs_new object containing tick mark ranges and lengths
}
\description{
make pretty breaks/tick marks for log scale ggplots
}
